fixed stars, inputs, some form styles
